Reduce the size of the gamma8 array
By bit-shifting the color values two bits to the right, we discard the low bits and have
only 64 values per color instead of 256. The trade-off is speed, unless the color values
are restricted to 6 bits at a higher level. We could save even more by cutting the size of
the array in half again, and possibly even storing an RGB structure in two bytes (using a
bit field) instead of three. 32 different brightness levels per color is probably
sufficient on something as coarse as a keyboard.